requirements.txt
pytest
pytest-selenium
termcolor
allure-python-commons
webdriver-manager